requirements/requirements.txt
Django==1.1.1
django-tagging==0.3.1
django_debug_toolbar==0.8.1
celery==1.0.0
-e git+http://github.com/ask/carrot.git#egg=carrot
-e git+http://github.com/ask/ghettoq.git#egg=ghettoq
poster==0.5.0
